+++ app/models/person.rb
@@ -1,6 +1,4 @@
 class Person < ApplicationRecord
-  include ReadyToApply
-
   # Attributes
 
   alias_attribute :owner, :main
@@ -13,31 +11,24 @@ class Person < ApplicationRecord
     owner ? "owner" : "companion"
   end
 
-  delegate :credit_score, :will_apply_for_loan,
-    :business_spending_usd, :has_business, :has_business?, :has_business_with_ein?,
-    :has_business_without_ein?, :no_business?,
-    to: :spending_info, allow_nil: true
-
   def onboarded_spending?
     !!spending_info&.persisted?
   end
 
   def onboarded?
     onboarded_eligibility? && onboarded_balances? && (
-      (ineligible?) || (
-        onboarded_cards? && onboarded_spending? && readiness_given?
-      )
+      (ineligible?) || (onboarded_cards? && onboarded_spending?)
     )
   end
 
   def can_receive_recommendations?
-    onboarded? && eligible? && ready_to_apply?
+    onboarded? && eligible? && ready?
   end
 
   def status
     if self.ineligible?
       "Ineligible"
-    elsif self.readiness_status.ready
+    elsif self.ready?
       "Ready"
     else
       "Eligible(NotReady)"
@@ -53,6 +44,17 @@ class Person < ApplicationRecord
       !eligible
     end
     alias_method :ineligible?, :ineligible
+  end
+
+  concerning :Readiness do
+    def onboarded_readiness?
+      !ready.nil?
+    end
+
+    def unready
+      !ready?
+    end
+    alias_method :unready?, :unready
   end
 
   # Validations