+++ spec/factories/accounts.rb
@@ -74,9 +74,6 @@ FactoryGirl.define do
       after(:build) do |acc|
         acc.people.each do |person|
           person.onboarded_cards = person.onboarded_balances = true
-          unless person.readiness_status.present?
-            person.build_readiness_status(ready: false)
-          end
         end
       end
     end
@@ -85,11 +82,7 @@ FactoryGirl.define do
       onboarded
       after(:build) do |acc|
         acc.people.each do |person|
-          if person.readiness_status.present?
-            person.readiness_status.ready = true
-          else
-            person.build_readiness_status(ready: true)
-          end
+          person.update_attribute(:ready, true)
         end
       end
     end