+++ spec/factories/readiness_status.rb
@@ -1,6 +1,0 @@
-FactoryGirl.define do
-  factory :readiness_status do
-    person
-    ready true
-  end
-end