+++ spec/models/person_spec.rb
@@ -58,12 +58,6 @@ describe Person do
         person.onboarded_cards = true
         expect(person).not_to be_onboarded
         person.onboarded_balances = true
-        expect(person).not_to be_onboarded
-        # Don't care if the user is ready:
-        person.build_readiness_status(ready: false)
-        allow(person.readiness_status).to receive(:persisted?) { true }
-        expect(person).to be_onboarded
-        person.readiness_status.ready = true
         expect(person).to be_onboarded
       end
     end