+++ CHANGELOG.md
@@ -2,10 +2,28 @@
 
 ## August 2016
 
+*   Remove 'readiness' page from survey; ask this question on the spending info
+    survey instead.
+
+    *George Millo*
+
+*   `track_intercom_event` RSpec matcher is now aliased as
+    `track_intercom_events` and can test that multiple Intercom events are
+    tracked at once.
+
+    *George Millo*
+
 *   Make all Form Objects use Virtus if they weren't already. Remove
     ApplicationForm#attr_boolean_accessor, as it's now no longer necessary.
 
     *George Millo*
+
+*   Remove ReadinessStatus model from project.
+    Add 'ready' and 'unreadiness_reason' to Person.
+
+    PT #127564851
+
+    *Ryan Vredenburg*
 
 *   Add tracking code for Heap